+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/) 
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## 0.8.2 (2018-10-XX)
+
+### Added
+
+- `VideoThumbGenerator::makeThumbnail()` now checks that output file exists and eventually throws `NoOutputGeneratedException`. [17](https://github.com/soluble-io/soluble-mediatools/issues/17)
+ 
+
 ## 0.8.1 (2018-10-25)
 
 ### Fixed